refactor(capsule): move capsule-builder to extensions/capsule
