+++ e2e/harmony/extensions-config.e2e.3.ts
@@ -7,7 +7,7 @@ const assertArrays = require('chai-arrays');
 
 chai.use(assertArrays);
 
-describe('harmony extension config', function() {
+describe.only('harmony extension config', function() {
   this.timeout(0);
   const helper = new Helper();
