--- src/environment/capsule-list.ts
@@ -1,4 +1,0 @@
-import BitIdAndValueArray from '../bit-id/bit-id-and-value-array';
-import { ComponentCapsule } from '../extensions/capsule-ext';
-
-export default class CapsuleList extends BitIdAndValueArray<ComponentCapsule> {}
