--- src/environment/capsule-paths.ts
@@ -1,4 +1,0 @@
-import { PathOsBasedAbsolute } from '../utils/path';
-import BitIdAndValueArray from '../bit-id/bit-id-and-value-array';
-
-export default class CapsulePaths extends BitIdAndValueArray<PathOsBasedAbsolute> {}
