+++ src/extensions/capsule/capsule-builder.ts
@@ -4,20 +4,20 @@ import os from 'os';
 import hash from 'object-hash';
 import v4 from 'uuid';
 import filenamify from 'filenamify';
-import { BitId } from '../bit-id';
-import orchestrator, { CapsuleOrchestrator } from '../extensions/capsule/orchestrator/orchestrator';
-import { PackageManager } from '../extensions/package-manager';
-import { CapsuleOptions, CreateOptions } from '../extensions/capsule/orchestrator/types';
-import Consumer from '../consumer/consumer';
-import { ComponentCapsule } from '../extensions/capsule-ext';
-import { getComponentLinks } from '../links/link-generator';
-import { Queue } from '../utils';
-import ManyComponentsWriter, { ManyComponentsWriterParams } from '../consumer/component-ops/many-components-writer';
-import { ComponentWithDependencies, loadScope } from '../scope';
-import { getManipulateDirForComponentWithDependencies } from '../consumer/component-ops/manipulate-dir';
-import Graph from '../scope/graph/graph';
-import Component from '../consumer/component';
-import { loadConsumerIfExist } from '../consumer';
+import { BitId } from '../../bit-id';
+import orchestrator, { CapsuleOrchestrator } from '../capsule/orchestrator/orchestrator';
+import { PackageManager } from '../package-manager';
+import { CapsuleOptions, CreateOptions } from '../capsule/orchestrator/types';
+import Consumer from '../../consumer/consumer';
+import { ComponentCapsule } from '../capsule-ext';
+import { getComponentLinks } from '../../links/link-generator';
+import { Queue } from '../../utils';
+import ManyComponentsWriter, { ManyComponentsWriterParams } from '../../consumer/component-ops/many-components-writer';
+import { ComponentWithDependencies, loadScope } from '../../scope';
+import { getManipulateDirForComponentWithDependencies } from '../../consumer/component-ops/manipulate-dir';
+import Graph from '../../scope/graph/graph';
+import Component from '../../consumer/component';
+import { loadConsumerIfExist } from '../../consumer';
 import CapsulePaths from './capsule-paths';
 import CapsuleList from './capsule-list';
