+++ src/extensions/capsule/capsule-paths.ts
@@ -1,0 +1,4 @@
+import { PathOsBasedAbsolute } from '../../utils/path';
+import BitIdAndValueArray from '../../bit-id/bit-id-and-value-array';
+
+export default class CapsulePaths extends BitIdAndValueArray<PathOsBasedAbsolute> {}
