+++ src/extensions/component/component.provider.ts
@@ -1,5 +1,5 @@
 import { ComponentFactory } from '../component';
-import Capsule from '../../environment/capsule-builder';
+import Capsule from '../capsule/capsule-builder';
 
 type ComponentDeps = [Capsule];
 type ComponentConfig = {};
